factor sas/x509 connectors into shared crate
