+++ identity/aziot-hub-client-async/Cargo.toml
@@ -20,8 +20,9 @@ serde = "1"
 serde_json = "1"
 url = "2"
 
-aziot-cert-common-http = { path = "../../cert/aziot-cert-common-http" }
 aziot-cert-client-async = { path = "../../cert/aziot-cert-client-async" }
+aziot-cert-common-http = { path = "../../cert/aziot-cert-common-http" }
+aziot-cloud-client-async-common = { path = "../aziot-cloud-client-async-common" }
 aziot-identity-common = { path = "../aziot-identity-common" }
 aziot-key-client = { path = "../../key/aziot-key-client" }
 aziot-key-client-async = { path = "../../key/aziot-key-client-async" }
